Extract verifyDataModel helper function
