+++ .gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 build/
 dist/
+.vscode
 node_modules/
 .prettierignore
 .env*
